feat: convert to typescript
